+++ src/components/LandingPages/News/index.js
@@ -8,17 +8,7 @@ import PresenterFactory from '../../APIInlinePresenterFactory'
 import * as statuses from '../../../constants/APIStatuses'
 import { flattenLocale } from '../../../shared/ContentfulLibs'
 
-export const sortNews = (left, right) => {
-  let a = new Date(left.fields.displayStartDate)
-  let b = new Date(right.fields.displayStartDate)
-
-  if (a < b) {
-    return 1
-  } else if (b < a) {
-    return -1
-  }
-  return 0
-}
+import { sortNews } from '../../Home/News'
 
 const mapStateToProps = (state) => {
   let allNews = []
@@ -29,11 +19,6 @@ const mapStateToProps = (state) => {
       .map((entry) => {
         flattenLocale(entry.fields, 'en-US')
         return entry
-      })
-      .filter((entry) => {
-        let start = new Date(entry.fields.displayStartDate)
-        let end = new Date(entry.fields.displayEndDate)
-        return start <= now && end >= now
       })
       .sort(sortNews)
   }